Clear the last x64 warnings and rewrite the readiness doc from measurement
Twelve left after the two sweeps. The size_t ones are the same
mechanical cast: fread/fwrite returns in Fbf and NetPlay, strlen into a
registry length and a NETMSG field, a reverse loop index in TextDraw
(cast before the -1, so an empty string still gives -1 rather than
wrapping through size_t).

Two were not narrowings. Display3D cast a BOOL message flag to
VIEW_PROXIMITY* purely to negate it, which is now just the negation.
Effects compared UDWORD against arithmetic on UWORD fields, which lands
in signed int; the operands now agree.

X64Readiness.md described this work as unmeasured and unstarted, which
it no longer is. The Watch section is replaced by what the builds
actually reported: 203 unique warnings on the first x64 build, and the
audit's expectation inverted -- 130 pointer truncations against 35
size_t narrowings, because the audit looked for pointers parked in
integers and the dominant idiom was the reverse. The three things the
audit did not anticipate at all, including the Multibot stack overrun,
are recorded there.

+++ NeuronClient/Fbf.cpp
@@ -120,7 +120,7 @@ int Neuron::FileGet(int fd)
 {
   if (fbf[fd].n == 0)
   {
-    fbf[fd].n = fread(fbf[fd].buffer, sizeof(int8), fbf[fd].buffersize, fbf[fd].fp);
+    fbf[fd].n = static_cast<int>(fread(fbf[fd].buffer, sizeof(int8), fbf[fd].buffersize, fbf[fd].fp));
     fbf[fd].b = fbf[fd].buffer;
   }
 
@@ -173,7 +173,7 @@ int Neuron::FilePut(int fd, int8 c)
 
   if (fbf[fd].n == fbf[fd].buffersize)
   {
-    i = fwrite(fbf[fd].buffer, sizeof(int8), fbf[fd].buffersize, fbf[fd].fp);
+    i = static_cast<int>(fwrite(fbf[fd].buffer, sizeof(int8), fbf[fd].buffersize, fbf[fd].fp));
     fbf[fd].n = 0;
     fbf[fd].b = fbf[fd].buffer;
   }
